requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.9.1
openai==1.63.0
supabase==2.15.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
faiss-cpu==1.11.0
python-dotenv==1.0.1
tiktoken==0.9.0